build: move ContractClient & AssembledTransaction
These are a bit higher-level and experimental, at this point, so let's
not clutter the global API or the bundle size unless people really want
it.
