+++ src/soroban/index.ts
@@ -9,7 +9,5 @@ export { Server, Durability } from './server';
 export { default as AxiosClient } from './axios';
 export { parseRawSimulation, parseRawEvents } from './parsers';
 export * from './transaction';
-export * from './contract_client';
-export * from './assembled_transaction';
 
 export default module.exports;
